Export and index changes, remove unused dependencies, relocate context to organisms
