+++ NaZdrowie/src/components/molecules/cards/CommentsCardForDoctor.tsx
@@ -1,5 +1,4 @@
 import { LinkButton, Comment } from "components/atoms";
-import primaryColors from "properties/colors";
 import { cardStyle, generalStyle } from "properties/styles";
 import { CommentData, CommentsCardProps } from "properties/types";
 import React from "react";
