+++ NaZdrowie/src/components/organisms/usePatientData.tsx
@@ -6,6 +6,7 @@ import {
   HealthFormFillOverlay,
   HealthFormResultOverlay,
 } from "components/molecules";
+import { useOverlay } from "components/organisms/context";
 import {
   CommentData,
   DoctorComment,
@@ -15,7 +16,6 @@ import {
   PatientResult,
   UserData,
 } from "properties/types";
-import { useOverlay } from "services/context";
 import {
   useFetchHealthComments,
   useFetchReferrals,
