+++ NaZdrowie/src/properties/types/InputProps.ts
@@ -1,4 +1,4 @@
 export type InputProps = {
   placeholder?: string;
-  onChange?: (value: (((prevState: string) => string) | string)) => void;
+  onChange?: (value: ((prevState: string) => string) | string) => void;
 };
