+++ NaZdrowie/src/properties/types/PatientDataProps.ts
@@ -38,8 +38,3 @@ export type ResultUpload = {
   testType: string;
   content: ResultDataContent;
 };
-
-export type UserData = {
-  id: number;
-  isDoctor: boolean;
-};
